Changed deck creation
Changed deck creation to follow the example given by the college board instructions. In the current code, it assumes that the same number of ranks, suits and values are going to be given, but that is not true. Furtheremore my code creates a card of a given rank and value for every possible suit.